Extacted filter creation logic from ScanUtils into FilterFactory.
